Safari-611.1.10/LayoutTests/editing/pasteboard/pasting-into-p-should-not-nest-p-expected.txt

This tests copying and pasting a p element into another p element doesn't end up in a nested p elements.

After paste:
| "\n"
| <p>
|   style="font-size: 12px; color: blue;"
|   <span>
|     id="source"
|     "Copy this line"
| "\n"
| <p>
|   style="font-size: 12px;"
|   <strong>
|     "Paste it below in the empty line below. The font size should be 12px and the text should be in blue."
| "\n"
| <p>
|   style="font-size: 12px; color: blue;"
|   <span>
|     id="source"
|     "Copy this line"
| <div>
|   <span>
|     <#selection-caret>
|     <br>
| "\n"

After innerHTML = innerHTML:
| "\n"
| <p>
|   style="font-size: 12px; color: blue;"
|   <span>
|     id="source"
|     "Copy this line"
| "\n"
| <p>
|   style="font-size: 12px;"
|   <strong>
|     "Paste it below in the empty line below. The font size should be 12px and the text should be in blue."
| "\n"
| <p>
|   style="font-size: 12px; color: blue;"
|   <span>
|     id="source"
|     "Copy this line"
| <#selection-caret>
| <div>
|   <span>
|     <br>
| "\n"
